feat(evals): typed runner core + automation/demo mode split
- pnpm evals = automation mode: no voiceover/narration policy anywhere
- pnpm fraimz = demo mode: unchanged behavior (drift check, narrated frames)
- runner converted to TypeScript on Node 24 native type-stripping (no build
  step, no new runtime deps); run.mjs stays as a thin bootstrap
- typed flow contract: defineFlow + FlowContext (evals/runner/flow.ts);
  new flows are *.flow.ts, legacy *.flow.mjs still load via compat shims
- scaffold emits typed .flow.ts stubs; plain stub without a script in
  automation mode, approved script still required in demo mode
- OPENWORK_EVAL_VOICEOVERS_DIR override for fixture isolation
- tsc gate (pnpm evals:typecheck) + runner unit tests (pnpm evals:test)
- nightly workflow pinned to pnpm fraimz --all to keep drift coverage
